initial work complete
More testing is needed but the tasks sensitive details are now masked when task event is emitted. Have to double check if these are the only tasks and properties that should have sensitive data masked. 
Also vitest is downgraded a bit. There is an issue with the vitest vscode extension that do not exists if vitest is bit lower version